+++ doc/feature to test table.md
@@ -20,9 +20,9 @@ Notice:
 |      2b      | Create data classes instead of using nested collection to store information about the cucumber feature               | N/A                                                                                                                                                                    | N/A                                                                                                                                                      | This is also done by designing the new structure |
 |      2c      | Relax the requirement where an int in the feature file is always mapping into a int in step def method int parameter | `scs.comp5903.cucumber.builder.JStepParameterExtractorTest#canIgnoreIfNotDeclared`                                                                                     | None                                                                                                                                                     |                                                  |
 |      2d      | Relax the requirement where step definition method name must be in a fixed pattern                                   | None                                                                                                                                                                   | `scs.comp5903.cucumber.sample.RummikubEasyCucumberIntegrationTest`                                                                                       |                                                  |
-|      4a      | Tags supports                                                                                                        | Any test method name began with "tag" in `scs.comp5903.cucumber.parser.LineByLineParserParsingTest`<br/>and `scs.comp5903.cucumber.execution.tag.BaseFilteringTagTest` | `scs.comp5903.cucumber.parser.JFeatureFileParserTest#canParseTags`<br/>and `scs.comp5903.cucumber.sample.TagFilteringTest`                               |                                                  |
-|      4b      | Supports comments and multi-line description                                                                         | `scs.comp5903.cucumber.parser.LineByLineParserParsingTest` <br/>and `scs.comp5903.cucumber.parser.LineByLineParserBuildingTest`                                        | `scs.comp5903.cucumber.parser.JFeatureFileParserTest#canIgnoreCommentsAndDescription`                                                                    |                                                  |
-|      4c      | Support keywords "And" and "But"                                                                                     | `scs.comp5903.cucumber.model.feature.jstep.AbstractJStepTest`                                                                                                                  | `scs.comp5903.cucumber.parser.DetailBuilderTest` <br/>and any tests that runs the Rummikub feature file in `src/test/resources/sample/jfeature/rummikub` |                                                  |
+|      4a      | Tags supports                                                                                                        | Any test method name began with "tag" in `scs.comp5903.cucumber.parser.jfeature.LineByLineParserParsingTest`<br/>and `scs.comp5903.cucumber.execution.tag.BaseFilteringTagTest` | `scs.comp5903.cucumber.parser.jfeature.JFeatureFileParserTest#canParseTags`<br/>and `scs.comp5903.cucumber.sample.TagFilteringTest`                               |                                                  |
+|      4b      | Supports comments and multi-line description                                                                         | `scs.comp5903.cucumber.parser.jfeature.LineByLineParserParsingTest` <br/>and `scs.comp5903.cucumber.parser.jfeature.LineByLineParserBuildingTest`                                        | `scs.comp5903.cucumber.parser.jfeature.JFeatureFileParserTest#canIgnoreCommentsAndDescription`                                                                    |                                                  |
+|      4c      | Support keywords "And" and "But"                                                                                     | `scs.comp5903.cucumber.model.feature.jstep.AbstractJStepTest`                                                                                                                  | `scs.comp5903.cucumber.parser.jfeature.DetailBuilderTest` <br/>and any tests that runs the Rummikub feature file in `src/test/resources/sample/jfeature/rummikub` |                                                  |
 |      4d      | Support parameter types of double, BigDecimal, BigInteger                                                            | `scs.comp5903.cucumber.builder.JStepParameterExtractorRegexTest` <br/>and `scs.comp5903.cucumber.builder.JStepParameterExtractorTest#testExtractParameters2`           | `scs.comp5903.cucumber.sample.DecimalAndBigIntTest`                                                                                                      |                                                  |
 |      4e      | Allow users to provide their own instance of the step definition class                                               | None                                                                                                                                                                   | `scs.comp5903.cucumber.sample.CmsSimpleTest`                                                                                                             |                                                  |
 |      4f      | Allow users to provide multiple step definition instances/classes                                                    | None                                                                                                                                                                   | `scs.comp5903.cucumber.sample.EasyCucumberApiTest`                                                                                                       |                                                  |
